+++ mobile/android/android-components/components/browser/engine-system/src/main/java/mozilla/components/browser/engine/system/SystemEngineView.kt
@@ -17,6 +17,7 @@ import android.util.AttributeSet
 import android.view.View
 import android.webkit.CookieManager
 import android.webkit.DownloadListener
+import android.webkit.JsResult
 import android.webkit.PermissionRequest
 import android.webkit.SslErrorHandler
 import android.webkit.ValueCallback
@@ -47,6 +48,7 @@ import mozilla.components.concept.engine.prompt.PromptRequest
 import mozilla.components.concept.engine.request.RequestInterceptor.InterceptionResponse
 import mozilla.components.support.ktx.android.content.isOSOnLowMemory
 import mozilla.components.support.utils.DownloadUtils
+import java.util.Date
 
 /**
  * WebView-based implementation of EngineView.
@@ -59,6 +61,9 @@ class SystemEngineView @JvmOverloads constructor(
 ) : FrameLayout(context, attrs, defStyleAttr), EngineView, View.OnLongClickListener {
 
     private var session: SystemEngineSession? = null
+    internal var jsAlertCount = 0
+    internal var shouldShowMoreDialogs = true
+    internal var lastDialogShownAt = Date()
 
     /**
      * Render the content of the given session.
@@ -123,6 +128,7 @@ class SystemEngineView @JvmOverloads constructor(
                     onNavigationStateChange(view.canGoBack(), view.canGoForward())
                 }
             }
+            resetJSAlertAbuseState()
         }
 
         override fun onPageFinished(view: WebView?, url: String?) {
@@ -300,6 +306,44 @@ class SystemEngineView @JvmOverloads constructor(
             session?.internalNotifyObservers { onContentPermissionRequest(SystemPermissionRequest(request)) }
         }
 
+        override fun onJsAlert(view: WebView, url: String?, message: String?, result: JsResult): Boolean {
+            val session = session ?: return super.onJsAlert(view, url, message, result)
+
+            // When an alert is triggered from a iframe, url is equals to about:blank, using currentUrl as a fallback.
+            val safeUrl = if (url.isNullOrBlank()) {
+                session.currentUrl
+            } else {
+                if (url.contains("about")) session.currentUrl else url
+            }
+
+            val title = context.getString(R.string.mozac_browser_engine_system_alert_title, safeUrl)
+
+            val onDismiss: () -> Unit = {
+                result.cancel()
+            }
+
+            if (shouldShowMoreDialogs) {
+
+                session.notifyObservers {
+                    onPromptRequest(
+                        PromptRequest.Alert(
+                            title,
+                            message ?: "",
+                            areDialogsBeingAbused(),
+                            onDismiss
+                        ) { shouldNotShowMoreDialogs ->
+                            shouldShowMoreDialogs = !shouldNotShowMoreDialogs
+                            result.confirm()
+                        })
+                }
+            } else {
+                result.cancel()
+            }
+
+            updateJSDialogAbusedState()
+            return true
+        }
+
         override fun onShowFileChooser(
             webView: WebView?,
             filePathCallback: ValueCallback<Array<Uri>>?,
@@ -449,7 +493,46 @@ class SystemEngineView @JvmOverloads constructor(
 
     override fun canScrollVerticallyDown() = session?.webView?.canScrollVertically(1) ?: false
 
+    private fun resetJSAlertAbuseState() {
+        jsAlertCount = 0
+        shouldShowMoreDialogs = true
+    }
+
+    internal fun updateJSDialogAbusedState() {
+        if (!areDialogsAbusedByTime()) {
+            jsAlertCount = 0
+        }
+        ++jsAlertCount
+        lastDialogShownAt = Date()
+    }
+
+    internal fun areDialogsBeingAbused(): Boolean {
+        return areDialogsAbusedByTime() || areDialogsAbusedByCount()
+    }
+
+    @Suppress("MagicNumber")
+    internal fun areDialogsAbusedByTime(): Boolean {
+        return if (jsAlertCount == 0) {
+            false
+        } else {
+            val now = Date()
+            val diffInSeconds = (now.time - lastDialogShownAt.time) / 1000 // 1 second equal to 1000 milliseconds
+            diffInSeconds < MAX_SUCCESSIVE_DIALOG_SECONDS_LIMIT
+        }
+    }
+
+    internal fun areDialogsAbusedByCount(): Boolean {
+        return jsAlertCount > MAX_SUCCESSIVE_DIALOG_COUNT
+    }
+
     companion object {
+
+        // Maximum number of successive dialogs before we prompt users to disable dialogs.
+        internal const val MAX_SUCCESSIVE_DIALOG_COUNT: Int = 2
+
+        // Minimum time required between dialogs in seconds before enabling the stop dialog.
+        internal const val MAX_SUCCESSIVE_DIALOG_SECONDS_LIMIT: Int = 3
+
         @Volatile
         internal var URL_MATCHER: UrlMatcher? = null